+++ packages/orq-rc/src/orq_ai_sdk/models/conversationresponse.py
@@ -70,7 +70,7 @@ class ConversationResponse(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kdw0p5pc4sny6s95kh84244q"
+        "conv_01kdwqh4ervq8mzdz4abycdtjq"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """