+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1166,7 +1166,7 @@ class Evaluations3(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-01-01T05:32:31.884Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-01-01T12:11:44.032Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1218,7 +1218,7 @@ class Evaluations2(BaseModel):
 
     source: Optional[EvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-01-01T05:32:31.884Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-01-01T12:11:44.032Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1270,7 +1270,7 @@ class Evaluations1(BaseModel):
 
     source: Optional[Source] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-01-01T05:32:31.883Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-01-01T12:11:44.032Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1349,5 +1349,5 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-01T05:32:17.721Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-01T12:11:30.915Z")
     r"""The date and time the resource was last updated"""