+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -239,7 +239,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KDW0P642AF6B0ZRPS9FE4MQF"
+        "01KDWQH4XFDA7WG3JYYPSAHVC4"
     )
     r"""The unique identifier of the data source"""