+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -101,5 +101,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-01-01T05:32:21.931Z")
+    created: Optional[datetime] = parse_datetime("2026-01-01T12:11:34.939Z")
     r"""The date and time the resource was created"""