+++ packages/orq-rc/src/orq_ai_sdk/models/updateconversationop.py
@@ -150,7 +150,7 @@ class UpdateConversationResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kdw0p5pkwz5ay0y32cmb0hrs"
+        "conv_01kdwqh4f051sga852qennxm22"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """