+++ packages/orq-rc/RELEASES.md
@@ -4489,3 +4489,13 @@ Based on:
 - [python v4.2.0-rc.4] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.4] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.4 - packages/orq-rc
+
+## 2026-01-01 12:14:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.680.11 (2.788.15) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.5] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.5] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.5 - packages/orq-rc